requirements.txt
# RAG Application Dependencies

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Streamlit UI
streamlit>=1.28.0

# Inngest Event-Driven
inngest>=0.2.0

# Vector Database
qdrant-client>=1.7.0

# Graph Database
neo4j>=5.0.0

# Medical NER (Transformers — no C++ build tools required)
transformers>=4.30.0
torch>=2.0.0

# Document Processing
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.0

# LLM and Embeddings
openai>=1.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0         # Required by pritamdeka/S-PubMedBert-MS-MARCO

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# HTTP Requests (for Streamlit app)
requests>=2.31.0

# Progress bar
tqdm>=4.66.0

# Testing
pytest>=7.4.0

datasets>=2.14.0

# PMC pipeline additions
tiktoken>=0.5.0                   # accurate token counting for PMCChunker
rapidfuzz>=3.0.0                  # used by EntityLinker
# Optional NER backends (scispacy requires C++ build tools — skip on Windows):
# gliner>=0.2.0

# Observability (Arize Phoenix + OpenTelemetry)
arize-phoenix[otel]>=4.0.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
grpcio>=1.50.0